Starter DMRIds.dat file
To get you started, here is a DMRIds.dat fle.  Place in /var/lib/mmdvm or the path your specified in P25Reflector.ini (Name=).  If you are running dvswitch-server this list will be in /var/lib/mmdvm and will be updated automatically.  If not, it can be obtained and updated manually from RadioID.  This script may be useful if you are not running dvs: https://github.com/g4klx/MMDVMHost/blob/master/linux/DMRIDUpdate.sh